fixed eval function

+++ basic_eval.h
@@ -2,6 +2,6 @@
 #define _BASIC_EVAL_H
 #include "boardstate.h"
 
-int basicEvaluation(BSTATE* currentboard);
+float basicEvaluation(BSTATE* currentboard);
 
 #endif
